Move json methods to op-service
