--- op-service/jsonutil/write.go
@@ -1,18 +1,0 @@
-package jsonutil
-
-import (
-	"encoding/json"
-	"os"
-)
-
-func WriteJSON(outfile string, input interface{}) error {
-	f, err := os.OpenFile(outfile, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o666)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	enc := json.NewEncoder(f)
-	enc.SetIndent("", "  ")
-	return enc.Encode(input)
-}
